requirements.txt
# FastAPI and API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Kafka
kafka-python==2.0.2

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
joblib==1.3.2

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Visualization
streamlit==1.28.1
plotly==5.18.0

# HTTP Requests
requests==2.31.0